# backend/swagger.yaml
openapi: 3.0.2
info:
  title: Items_MS API
  version: 1.0.0
  description: List of items API
  contact:
    name: Maarika Lukk
    url: 'https://github.com/maarikal'
    email: maarika.lukk@voco.ee
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
paths:
  /items:
    summary: Path used to manage the list of items.
    description: >-
      The REST endpoint/path used to list and create zero or more `items` entities.  This path contains
      a `GET` and `POST` operation to perform the list and create tasks, respectively.
    get:
      tags:
        - Items
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Items returned:
                  value:
                    description: some text
                    image: a picture
                    name: new namet
            text/xml:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Item returned:
                  value: "<Item>\r\n    <description>some text</description>\r\n    <image>returned item image</image>\r\n    <name>returned item name</name>\r\n</Item>"
          description: Items returned
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Unauthorized:
                  value:
                    error: Unauthorized
                Bearer:
                  value:
                    error: Authorization header is not in Bearer XXX format
                Session not found:
                  value:
                    error: Session not found
                User not found:
                  value:
                    error: User not found
          description: Unauthorized
        '406':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Not Acceptable:
                  value:
                    error: Not acceptable
          description: Not accetable
      security:
        - Bearer: [ ]
      summary: Get all items
      description: Gets a list of all `items` entities.
    post:
      requestBody:
        description: A new `items` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
            examples:
              ' PostItemsRequestBody_json':
                value:
                  name: new bird
                  description: it is green and small
                  image: >-
                    https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhBY9h14uKCSgCxNCZEhClbX2ksq2Fi6iCmQ&usqp=CAU
          text/xml:
            schema:
              $ref: '#/components/schemas/Item'
            examples:
              PostItemsRequestBody_xml:
                value: "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<root>\r\n\t<name>item nr 1</name>\r\n\t<description>it is green and small</description>\r\n\t<image>https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhBY9h14uKCSgCxNCZEhClbX2ksq2Fi6iCmQ&amp;usqp=CAU</image>\r\n</root>\r\n"
        required: true
      tags:
        - Items
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Item returned:
                  value:
                    description: some text
                    image: returned item image
                    name: returned item name
            text/xml:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Item returned:
                  value: "<head><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<root>\r\n\t<name>item nr 1</name>\r\n\t<description>it is green and small</description>\r\n\t<image>https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhBY9h14uKCSgCxNCZEhClbX2ksq2Fi6iCmQ&amp;usqp=CAU</image>\r\n</root>\r\n</head>"
          description: Return item
        '406':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Not Acceptable:
                  value:
                    error: Not Acceptable
          description: Not Acceptable
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Internal Server Error:
                  value:
                    Error: Internal Server Error
          description: Internal Server Error
      security:
        - Bearer: [ ]
      operationId: createitems
      summary: Create a item
      description: Creates a new instance of a `items`.
...
  '/items?id={itemsId}':
    summary: Path used to manage a single items.
    description: >-
      The REST endpoint/path used to get, update, and delete single instances of an `items`.  This path
      contains `GET`, `PUT`, and `DELETE` operations used to perform the get, update, and delete tasks,
      respectively.
    post:
      requestBody:
        description: Added item information.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
            examples:
              PostItemIdRequestBody_json:
                value:
                  name: new bird2
                  description: it is green and small
                  image: >-
                    https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhBY9h14uKCSgCxNCZEhClbX2ksq2Fi6iCmQ&usqp=CAU
          text/xml:
            schema:
              $ref: '#/components/schemas/Item'
            examples:
              PostItemIdRequestBody_xml:
                value: "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<root>\n\t<name>item nr 2</name>\n\t<description>it is green and small</description>\n\t<image>https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhBY9h14uKCSgCxNCZEhClbX2ksq2Fi6iCmQ&amp;usqp=CAU</image>\n</root>\n"
        required: true
      tags:
        - Items
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Created:
                  value:
                    description: some text
                    image: one image
                    name: some name
            text/xml:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Created:
                  value: "<head><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<Item>\r\n    <id></id>\r\n    <description>some text</description>\r\n    <image>returned item image</image>\r\n    <name>returned item name</name>\r\n</Item>\r\n</head>"
          description: Item created
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                ' Successful response.':
                  value:
                    Item: Successful response
            text/xml:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                ' Successful response.':
                  value: <text>Successful response</text>
          description: Successful response.
        '406':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Not Acceptable:
                  value:
                    error: Not Acceptable
          description: Not Acceptable
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Internal Server Error:
                  value:
                    Error: Internal Server Error
          description: "\t\r\nInternal Server Error"
      security:
        - Bearer: [ ]
      operationId: addeditem
      summary: Add a new item
      description: Added new item.
    delete:
      tags:
        - Items
      responses:
        '204':
          description: Successful response.
      security:
        - Bearer: [ ]
      operationId: deleteitems
      summary: Delete a item
      description: Deletes an existing `items`.
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
            examples:
              PatchItemRequestBody_json:
                value:
                  name: new bird
                  description: it is green and small
                  image: >-
                    https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhBY9h14uKCSgCxNCZEhClbX2ksq2Fi6iCmQ&usqp=CAU
          text/xml:
            schema:
              $ref: '#/components/schemas/Item'
            examples:
              PatchItemRequestBody_xml:
                value: "<head><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<Item>\r\n    <id></id>\r\n    <description>some text</description>\r\n    <image>returned item image</image>\r\n    <name>returned item name</name>\r\n</Item>\r\n</head>"
        required: true
      tags:
        - Items
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Item created:
                  value:
                    description: some text
                    image: some image
                    name: some name
            text/xml:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Item created:
                  value: "<head><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<Item>\r\n    <id></id>\r\n    <description>some text</description>\r\n    <image>returned item image</image>\r\n    <name>returned item name</name>\r\n</Item>\r\n</head>"
          description: Item created
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Response:
                  value:
                    description: some text
                    image: some image
                    name: some name
            text/xml:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Successful response:
                  value: "<Item>\r\n    <id></id>\r\n    <description>some text</description>\r\n    <image>returned item image</image>\r\n    <name>returned item name</name>\r\n</Item>"
          description: Successful response.
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Internal Server Error:
                  value:
                    Error: Internal Server Error
          description: Internal Server Error
      security:
        - Bearer: [ ]
      operationId: updateitem
      summary: Update a item
      description: Updates an existing item.
    parameters:
      - name: itemsId
        description: A unique identifier for a `items`.
        schema:
          type: string
        in: path
        required: true
  /oAuth2Login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Session'
            examples:
              GoogleAuthRequestBody:
                value:
                  token: >-
                    eyJhbGciOiJSUzI1NiIsImtpZCI6IjA1MTUwYTEzMjBiOTM5NWIwNTcxNjg3NzM3NjkyODUwOWJhYjQ0YWMiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20iLCJuYmYiOjE2ODY4MzMyMDMsImF1ZCI6IjY2ODI1MDMwMTcwNC1xN2o0dDh0bmttazg4ajNkNmpzcmt1anQ3NDMxMXVuYi5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsInN1YiI6IjEwMjI1OTAzMjY0MjI1NjgwMDU1MCIsImVtYWlsIjoic2lsdmVyLnNpcnBAZ21haWwuY29tIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsImF6cCI6IjY2ODI1MDMwMTcwNC1xN2o0dDh0bmttazg4ajNkNmpzcmt1anQ3NDMxMXVuYi5hcHBzLmdvb2dsZXVzZXJjb250ZW50LmNvbSIsIm5hbWUiOiJTaWx2ZXIgU2lycCIsInBpY3R1cmUiOiJodHRwczovL2xoMy5nb29nbGV1c2VyY29udGVudC5jb20vYS9BQWNIVHRlWnJUc1FzRkNlN0ZpbFNwelFoZXFReGpSTnJ6cGFoaUhjRFlfZz1zOTYtYyIsImdpdmVuX25hbWUiOiJTaWx2ZXIiLCJmYW1pbHlfbmFtZSI6IlNpcnAiLCJpYXQiOjE2ODY4MzM1MDMsImV4cCI6MTY4NjgzNzEwMywianRpIjoiYjgwODlmZjc1ZmRkNTQ3ODQ2Njk5ZTc1NzVjMjIxNWFiMWM3Y2UxZSJ9.xBc_fkkf3MRZMjSZN9IxQTUHaeAo5B6WJL-ZBvyEMknregEtk7m9KepEZmXYasEeFUh0KRq-MEz_037K34qSW5Pv2Hce5cEpY3pRSy0mvrWsxTUQ9LptkRjGnXl2tE1zU26NkX77d75Gn7B0vSai7uNNhdRZLNsyytkNJQvXJ28v18lmyLvTT9vilIjGOqFN1ZBrCv5tUCfGy2e6X8KTtEUJzmLWkrHSJo5sWmvdUedicpuaJCMFl65mTMKlb2PsFeQDQousOVAxgCJgtCFIibhAZ7lHqWZ0VUt-6GKv81AhyJqwXY7YbrlfyewVYAnBkuzzszmU-erauY3qLnBfXQ
        required: true
      tags:
        - oAuth2Login
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              examples:
                SessionId:
                  value:
                    user: test@test.ee
                    password: qwerty123
          description: Returns sessionId
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Bad Request:
                  value:
                    error: Bad Request
          description: Bad Request
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Internal Server Error:
                  value:
                    Error: Internal Server Error
          description: Internal Server Error
      security:
        - Bearer: [ ]
  /logs:
    get:
      tags:
        - Logs
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
              examples:
                Logs sent:
                  value:
                    logs:
                      - item:
                          description: >-
                            Emeepabi topha edmuk wujzi hubdi jo riwbeoc ozgi caf
                            celurbe gubi neg fopuvfu kiakba rowdelwol la hapupes.
                          id: 353
                          image: 'https://picsum.photos/96/96?random=447'
                          name: uuuuuus2
                        message: ITEM UPDATED
                        timestamp: '01-06-2023 11:12:51'
          description: |
            Logs sent
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                Unauthorized:
                  value:
                    error: Unauthorized
                Bearer:
                  value:
                    error: Authorization header is not in Bearer XXX format
                Session not found:
                  value:
                    error: Session not found
                User not found:
                  value:
                    error: User not found
          description: "\t\r\nUnauthorized"
      security:
        - Bearer: [ ]
components:
  schemas:
...
    Item:
      title: Root Type for items
      description: Items list
      type: object
      properties:
        description:
          type: string
        image:
          type: string
        name:
          type: string
      example:
        name: item nr 1
        description: it is green and small
        image: >-
          https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRhBY9h14uKCSgCxNCZEhClbX2ksq2Fi6iCmQ&usqp=CAU
    Session:
      title: Root Type for Session
      description: 'If user is on webpage and have logged in, then he/she has a sessionId.'
      type: object
      properties:
        user:
          type: string
        password:
          type: string
      example:
        user: test@test.ee
        password: qwerty123
...
    Error:
      description: |-
        {
            "error:" "Error message"
        }
      type: object
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
...